backend/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
aiogram
python-dotenv
pydantic-settings